@@ -1,45 +1,46 @@
-//Plugin do Android
-apply plugin: 'com.android.application'
-//Plugin da liguagem Kotlin
-apply plugin: 'kotlin-android'
-//Adiciona funcionalidades estendida ao Kotlin
-apply plugin: 'kotlin-android-extensions'
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+    id 'kotlin-android-extensions'
+}
 
 android {
-    //Versão de complilação do app
-    compileSdkVersion 26
+    compileSdkVersion 30
+    buildToolsVersion "30.0.2"
+
     defaultConfig {
-        //Nome do pacote do projeto, deve ser único
-        applicationId "tictactoegame.com.br.tictactoe"
-        //Versão minima suportada
-        minSdkVersion 15
-        //Vesão alvo do aplicativo
-        targetSdkVersion 26
-        //Número da versão do código
+        applicationId "io.github.leoallvez.tictactoe"
+        minSdkVersion 16
+        targetSdkVersion 30
         versionCode 1
-        //Nomeclátura da versão do código
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
 }
-//dependências do projeto
+
 dependencies {
-    //inclui a pasta libs na compilação
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    //biblioteca do Kotlin
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    //biblioteca de compatibilidade do google
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    //biblioteca do constraint layout
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    //bibliotecas de testes
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation 'androidx.core:core-ktx:1.3.2'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    testImplementation 'junit:junit:4.+'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 }
