@@ -1,18 +1,14 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    //Versão do Kotlin
-    ext.kotlin_version = '1.3.10'
+    ext.kotlin_version = "1.4.31"
     repositories {
-        //Repositórios de dependências
         google()
         jcenter()
     }
     dependencies {
-        //Versão do plugin do gradle
-        classpath 'com.android.tools.build:gradle:3.5.0'
-        //Configuração do Kotlin nos projetos
+        classpath "com.android.tools.build:gradle:4.1.2"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
